requirements.txt
flask-restful
flask-cors
pytest
psycopg2-binary
pyscopg2